feat(landing): extract CheckoutPanel + account Sync section
